+++ src/ps.h
@@ -50,6 +50,8 @@ SEXP ps__cpu_count_logical();
 SEXP ps__cpu_count_physical();
 SEXP ps__users();
 SEXP ps__tty_size();
+SEXP ps__disk_partitions(SEXP all);
+SEXP ps__disk_usage(SEXP paths);
 
 /* Generic utils used from R */