+++ src/ps.h
@@ -52,6 +52,8 @@ SEXP ps__users();
 SEXP ps__tty_size();
 SEXP ps__disk_partitions(SEXP all);
 SEXP ps__disk_usage(SEXP paths);
+SEXP ps__system_memory();
+SEXP ps__system_swap();
 
 /* Generic utils used from R */