+++ src/arch/windows/process_handles.h
@@ -1,8 +1,3 @@
-/*
- * Copyright (c) 2009, Jay Loden, Giampaolo Rodola'. All rights reserved.
- * Use of this source code is governed by a BSD-style license that can be
- * found in the LICENSE file.
- */
 
 #ifndef __PROCESS_HANDLES_H__
 #define __PROCESS_HANDLES_H__