+++ src/arch/windows/process_info.c
@@ -1,11 +1,3 @@
-/*
- * Copyright (c) 2009, Jay Loden, Giampaolo Rodola'. All rights reserved.
- * Use of this source code is governed by a BSD-style license that can be
- * found in the LICENSE file.
- *
- * Helper functions related to fetching process information. Used by
- * _psutil_windows module methods.
- */
 
 #include <windows.h>
 #include <Psapi.h>