+++ src/common.c
@@ -1,10 +1,3 @@
-/*
- * Copyright (c) 2009, Giampaolo Rodola'. All rights reserved.
- * Use of this source code is governed by a BSD-style license that can be
- * found in the LICENSE file.
- *
- * Routines common to all platforms.
- */
 
 #ifndef _GNU_SOURCE
 #define _GNU_SOURCE 1