+++ src/windows.c
@@ -1,10 +1,3 @@
-/*
- * Copyright (c) 2009, Jay Loden, Giampaolo Rodola'. All rights reserved.
- * Use of this source code is governed by a BSD-style license that can be
- * found in the LICENSE file.
- *
- * Windows platform-specific module methods for _psutil_windows
- */
 
 // Fixes clash between winsock2.h and windows.h
 #define WIN32_LEAN_AND_MEAN