+++ NEWS.md
@@ -1,4 +1,6 @@
 
+# ps development version
+
 # ps 1.3.4
 
 * `ps_cpu_count()` now reports the correct number on Windows, even if