+++ NEWS.md
@@ -1,6 +1,8 @@
 
 # ps development version
 
+* New `ps_tty_size()` function to query the size of the terminal.
+
 # ps 1.3.4
 
 * `ps_cpu_count()` now reports the correct number on Windows, even if