+++ NEWS.md
@@ -1,5 +1,5 @@
 
-# ps development version
+# ps 1.4.0
 
 * ps is now under the MIT license.